pyproject.toml
[project]
name = "parwanweb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "JahangirMoro", email = "jahangeerrathore@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "email-validator>=2.2.0",
    "fastapi>=0.115.14",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.35.0",
]

[project.scripts]
parwanweb = "parwanweb:main"
parwanmain = "app.main:main"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
